added documentation

+++ main.py
@@ -18,7 +18,7 @@ if __name__ == '__main__':
     game_info['max_steps_per_episode'] = int(game_info['max_steps_per_episode'])
     game_info['training_phase'] = int(game_info['training_phase'])
     game_info['play_phase'] = int(game_info['play_phase'])
-    game_info['reward_per_coins'] = int(game_info['reward_per_coins'])
+    game_info['reward_per_coins'] = float(game_info['reward_per_coins'])
     if game_info['algorithm'] == "Q_Learning":
         game_info = q_learning_agent.main(game_info)
     else:
